requirements.txt
Django==1.7.1
psycopg2==2.5.1
django-vanilla-views==1.0.2
django-crispy-forms==1.4.0
django-admin-bootstrapped==1.6.3
django-codemirror-widget==0.4.0
PyYAML==3.11
gevent==1.0.1
gevent-psycopg2==0.0.3
django-annoying==0.8.0
